Refactor `handlePageFuntion`, add pagination search
